Move to stdint: model and reader

+++ src/model.c
@@ -24,8 +24,10 @@
  * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
  * POSSIBILITY OF SUCH DAMAGE.
  */
+#include <inttypes.h>
 #include <stdbool.h>
 #include <stddef.h>
+#include <stdint.h>
 #include <stdlib.h>
 #include <stdio.h>
 #include <string.h>
@@ -120,8 +122,8 @@ void mdl_free(mdl_t *mdl) {
  *   This reduce the risk of mistakes.
  */
 void mdl_sync(mdl_t *mdl) {
-	const size_t Y = qrk_count(mdl->reader->lbl);
-	const size_t O = qrk_count(mdl->reader->obs);
+	const uint64_t Y = qrk_count(mdl->reader->lbl);
+	const uint64_t O = qrk_count(mdl->reader->obs);
 	// If model is already synchronized, do nothing and just return
 	if (mdl->nlbl == Y && mdl->nobs == O)
 		return;
@@ -131,8 +133,8 @@ void mdl_sync(mdl_t *mdl) {
 	// case we also display a warning as this is probably not expected by
 	// the user. If only new observations was added, we will try to expand
 	// the model.
-	size_t oldF = mdl->nftr;
-	size_t oldO = mdl->nobs;
+	uint64_t oldF = mdl->nftr;
+	uint64_t oldO = mdl->nobs;
 	if (mdl->nlbl != Y && mdl->nlbl != 0) {
 		warning("labels count changed, discarding the model");
 		free(mdl->kind);  mdl->kind  = NULL;
@@ -148,16 +150,16 @@ void mdl_sync(mdl_t *mdl) {
 	mdl->nobs = O;
 	// Allocate the observations datastructure. If the model is empty or
 	// discarded, a new one iscreated, else the old one is expanded.
-	char   *kind = xrealloc(mdl->kind, sizeof(char  ) * O);
-	size_t *uoff = xrealloc(mdl->uoff, sizeof(size_t) * O);
-	size_t *boff = xrealloc(mdl->boff, sizeof(size_t) * O);
+	char     *kind = xrealloc(mdl->kind, sizeof(char    ) * O);
+	uint64_t *uoff = xrealloc(mdl->uoff, sizeof(uint64_t) * O);
+	uint64_t *boff = xrealloc(mdl->boff, sizeof(uint64_t) * O);
 	mdl->kind = kind;
 	mdl->uoff = uoff;
 	mdl->boff = boff;
 	// Now, we can setup the features. For each new observations we fill the
 	// kind and offsets arrays and count total number of features as well.
-	size_t F = oldF;
-	for (size_t o = oldO; o < O; o++) {
+	uint64_t F = oldF;
+	for (uint64_t o = oldO; o < O; o++) {
 		const char *obs = qrk_id2str(mdl->reader->obs, o);
 		switch (obs[0]) {
 			case 'u': kind[o] = 1; break;
@@ -176,14 +178,14 @@ void mdl_sync(mdl_t *mdl) {
 	// have to allocate a new vector and copy old values ourself.
 	if (oldF != 0) {
 		double *new = xvm_new(F);
-		for (size_t f = 0; f < oldF; f++)
+		for (uint64_t f = 0; f < oldF; f++)
 			new[f] = mdl->theta[f];
 		xvm_free(mdl->theta);
 		mdl->theta = new;
 	} else {
 		mdl->theta = xvm_new(F);
 	}
-	for (size_t f = oldF; f < F; f++)
+	for (uint64_t f = oldF; f < F; f++)
 		mdl->theta[f] = 0.0;
 	// And lock the databases
 	qrk_lock(mdl->reader->lbl, true);
@@ -197,7 +199,7 @@ void mdl_sync(mdl_t *mdl) {
  *   and labeling.
  */
 void mdl_compact(mdl_t *mdl) {
-	const size_t Y = mdl->nlbl;
+	const uint64_t Y = mdl->nlbl;
 	// We first build the new observation list with only observations which
 	// lead to at least one active feature. At the same time we build the
 	// translation table which map the new observations index to the old
@@ -205,29 +207,29 @@ void mdl_compact(mdl_t *mdl) {
 	info("    - Scan the model\n");
 	qrk_t *old_obs = mdl->reader->obs;
 	qrk_t *new_obs = qrk_new();
-	size_t *trans = xmalloc(sizeof(size_t) * mdl->nobs);
-	for (size_t oldo = 0; oldo < mdl->nobs; oldo++) {
+	uint64_t *trans = xmalloc(sizeof(uint64_t) * mdl->nobs);
+	for (uint64_t oldo = 0; oldo < mdl->nobs; oldo++) {
 		bool active = false;
 		if (mdl->kind[oldo] & 1)
-			for (size_t y = 0; y < Y; y++)
+			for (uint64_t y = 0; y < Y; y++)
 				if (mdl->theta[mdl->uoff[oldo] + y] != 0.0)
 					active = true;
 		if (mdl->kind[oldo] & 2)
-			for (size_t d = 0; d < Y * Y; d++)
+			for (uint64_t d = 0; d < Y * Y; d++)
 				if (mdl->theta[mdl->boff[oldo] + d] != 0.0)
 					active = true;
 		if (!active)
 			continue;
-		const char   *str  = qrk_id2str(old_obs, oldo);
-		const size_t  newo = qrk_str2id(new_obs, str);
+		const char     *str  = qrk_id2str(old_obs, oldo);
+		const uint64_t  newo = qrk_str2id(new_obs, str);
 		trans[newo] = oldo;
 	}
 	mdl->reader->obs = new_obs;
 	// Now we save the old model features informations and build a new one
 	// corresponding to the compacted model.
-	size_t *old_uoff  = mdl->uoff;  mdl->uoff  = NULL;
-	size_t *old_boff  = mdl->boff;  mdl->boff  = NULL;
-	double *old_theta = mdl->theta; mdl->theta = NULL;
+	uint64_t *old_uoff  = mdl->uoff;  mdl->uoff  = NULL;
+	uint64_t *old_boff  = mdl->boff;  mdl->boff  = NULL;
+	double   *old_theta = mdl->theta; mdl->theta = NULL;
 	free(mdl->kind);
 	mdl->kind = NULL;
 	mdl->nlbl = mdl->nobs = mdl->nftr = 0;
@@ -235,18 +237,18 @@ void mdl_compact(mdl_t *mdl) {
 	// The model is now ready, so we copy in it the features weights from
 	// the old model for observations we have kept.
 	info("    - Compact it\n");
-	for (size_t newo = 0; newo < mdl->nobs; newo++) {
-		const size_t oldo = trans[newo];
+	for (uint64_t newo = 0; newo < mdl->nobs; newo++) {
+		const uint64_t oldo = trans[newo];
 		if (mdl->kind[newo] & 1) {
 			double *src = old_theta  + old_uoff[oldo];
 			double *dst = mdl->theta + mdl->uoff[newo];
-			for (size_t y = 0; y < Y; y++)
+			for (uint64_t y = 0; y < Y; y++)
 				dst[y] = src[y];
 		}
 		if (mdl->kind[newo] & 2) {
 			double *src = old_theta  + old_boff[oldo];
 			double *dst = mdl->theta + mdl->boff[newo];
-			for (size_t d = 0; d < Y * Y; d++)
+			for (uint64_t d = 0; d < Y * Y; d++)
 				dst[d] = src[d];
 		}
 	}
@@ -262,15 +264,15 @@ void mdl_compact(mdl_t *mdl) {
  *   Save a model to be restored later in a platform independant way.
  */
 void mdl_save(mdl_t *mdl, FILE *file) {
-	size_t nact = 0;
-	for (size_t f = 0; f < mdl->nftr; f++)
+	uint64_t nact = 0;
+	for (uint64_t f = 0; f < mdl->nftr; f++)
 		if (mdl->theta[f] != 0.0)
 			nact++;
-	fprintf(file, "#mdl#%zu\n", nact);
+	fprintf(file, "#mdl#%"PRIu64"\n", nact);
 	rdr_save(mdl->reader, file);
-	for (size_t f = 0; f < mdl->nftr; f++)
+	for (uint64_t f = 0; f < mdl->nftr; f++)
 		if (mdl->theta[f] != 0.0)
-			fprintf(file, "%zu=%la\n", f, mdl->theta[f]);
+			fprintf(file, "%"PRIu64"=%la\n", f, mdl->theta[f]);
 }
 
 /* mdl_load:
@@ -280,15 +282,15 @@ void mdl_save(mdl_t *mdl, FILE *file) {
  */
 void mdl_load(mdl_t *mdl, FILE *file) {
 	const char *err = "invalid model format";
-	size_t nact = 0;
-	if (fscanf(file, "#mdl#%zu\n", &nact) != 1)
+	uint64_t nact = 0;
+	if (fscanf(file, "#mdl#%"SCNu64"\n", &nact) != 1)
 		fatal(err);
 	rdr_load(mdl->reader, file);
 	mdl_sync(mdl);
-	for (size_t i = 0; i < nact; i++) {
-		size_t f;
+	for (uint64_t i = 0; i < nact; i++) {
+		uint64_t f;
 		double v;
-		if (fscanf(file, "%zu=%la\n", &f, &v) != 2)
+		if (fscanf(file, "%"SCNu64"=%la\n", &f, &v) != 2)
 			fatal(err);
 		mdl->theta[f] = v;
 	}
